Make btree item, transaction and constructor for NewEventHandler and fully comment them all
