+++ README.md
@@ -61,3 +61,70 @@ Participation in the Kubernetes community is governed by the [Kubernetes Code of
 
 
 TODO add some lines about what to find ./hack
+
+### Code walkthrough
+
+#### The KPNG kube ... command path
+
+```go
+kube --kubeconfig=/var/lib/kpng/kubeconfig.conf to-local to-nft
+```
+
+- `kpng` is the kpng binary.
+- `kube` is the command which kpng runs, which watches and internalizes the Kubernetes network API (services and pods)
+- `to-local` is a cobra command, which proxies to a "local" backend (like, NFT).
+- `to-nft` is the name of the local backend, and is a **subcommand** of `to-local`. 
+
+Thus, KPNG uses a core command "kube" to start up, and then decides wether it is going to write iptables/nft/ebpf... rule
+based on the value of the subcommands for the `to-local` option which it recieves.
+
+The way that the *to-nft* subcommand is built up at runtime, when KPNG introspects and imports several pacages.
+
+- On startup, before parsing arguments, KPNG imports several *backend* packages
+- these imports result in `init()` functions being called,
+- the `init()` functions in those backends then go on to call the `Register()` function in the `backendcmd` pacakge.
+```go
+    // scorecmds_linux.go calls the `init()` functions in all the following packages... 
+    import (
+    _ "sigs.k8s.io/kpng/backends/ebpf"
+    _ "sigs.k8s.io/kpng/backends/iptables"
+    _ "sigs.k8s.io/kpng/backends/ipvs-as-sink"
+    _ "sigs.k8s.io/kpng/backends/nft"
+    _ "sigs.k8s.io/kpng/backends/userspacelin"
+    )
+```
+- Using the **cobra** API, KPNG now reads through the `Registered()` backends (which are called above), and makes command line options for these one at a time.  The logic for this
+is in the `ToLocalCmd(...)` function, 
+  - ToLocal: Connecting the changes in the k8s networking statespace to a backend, that implements the KPNG sink interface, which we'll go over next.
+
+
+In the code, the way that these subcommands are built up, is as follows:
+
+```mermaid
+sequenceDiagram
+user ->> "cmd/kpng/kube2store": "kpng +kubeconfig ...+to-nft..."
+
+    "cmd/kpng/storecmds" --> "nft" : import
+    "nft" ->> "cli/backendcmd": Register
+ 
+
+
+    "kube2store" ->> "commandbuilder":Commands 
+    "commandbuilder" --> "kube2store":"ToAPI,ToFile,ToLocal"
+
+    "commandbuilder" ->> "cli/backendcmd": Registered()
+```
+
+At this point, you now have KPNG watching the Kubernetes APIServer, and writing out new backend rules for every new service or pod that is created.
+
+Since the command line invocation above specifies 2 things:
+- Watch the kubernetes API Server
+- Send its events, decoded as KPNG objects to the NFT backend which will write routing rules
+
+We will now go through these 2 separate KPNG process which were just started.
+
+#### The "Kube" or "Kube2Store" command
+
+The Kube command is self explanatory.  It isnt dependent on any way, of the underlying backend (or lack thereof).  
+
+#### The "to-local to-nft" command
